+++ src/components/CodeBlock/prismjs-components/prism-components.ts
@@ -1,3 +1,6 @@
 declare module "*" {
     function registerLanguage(Prism: object): void;
+
+    export class Default {
+    }
 }