+++ src/components/Notifications/NotificationsItem/NotificationItem.styled.tsx
@@ -86,7 +86,5 @@ export const NotificationsItemStyled = styled.div`
         line-height : 0;
     }
 
-    &.dismissed {
-        transform: translateX(560px);
-    }
+    &.dismissed { transform : translateX(560px); }
 `;